+++ lib/src/chart_model.dart
@@ -6,7 +6,8 @@ library highcharts.src.chart_model;
 enum SeriesStyle { 
   name, color, colorByPoint, borderRadius, fillOpacity,
   size, innerSize, lineColor, dataLabels, sliced, selected,
-  markerEnabled, markerLineColor, markerSymbol, markerFillColor, markerLineWidth // marker
+  markerEnabled, markerLineColor, markerSymbol, markerFillColor, markerLineWidth, // marker
+  showInLegend,
 }
 
 abstract class ChartModel<S extends Comparable, C extends Comparable> {